+++ tracer/src/Datadog.Trace/Iast/Aspects/System.Data.Common/DbCommandAspect.cs
@@ -10,29 +10,11 @@ using Datadog.Trace.Iast.Dataflow;
 namespace Datadog.Trace.Iast.Aspects;
 
 /// <summary> DbCommandAspect class aspect </summary>
-[AspectClass("Microsoft.EntityFrameworkCore.Relational")]
+[AspectClass("System.Data,System.Data.Common")]
 [global::System.ComponentModel.Browsable(false)]
 [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
 public class DbCommandAspect
 {
-    /// <summary>
-    /// ReviewSqlString aspect
-    /// </summary>
-    /// <param name="command"> the DbCommand </param>
-    /// <returns> resulting sql query </returns>
-    [AspectMethodInsertBefore("System.Data.Entity.Core.EntityClient.EntityCommand::ExecuteReader(System.Data.CommandBehavior)", 1)]
-    [AspectMethodInsertBefore("System.Data.Entity.Core.EntityClient.EntityCommand::ExecuteReaderAsync(System.Data.CommandBehavior,System.Threading.CancellationToken)", 2)]
-    [AspectMethodInsertBefore("System.Data.Entity.Core.EntityClient.EntityCommand::ExecuteReaderAsync(System.Data.CommandBehavior)", 1)]
-    public static object ReviewSqlCommand(object command)
-    {
-        if (command is DbCommand dbCommand)
-        {
-            IastModule.OnSqlQuery(dbCommand.CommandText, IntegrationId.SqlClient);
-        }
-
-        return command;
-    }
-
     /// <summary>
     /// ReviewSqlString aspect
     /// </summary>
